requirements.txt
django
djangorestframework
django-oauth-toolkit
django-cors-middleware
